// ERP.Web/src/app/theme/pages/models/driverinformation.ts
export class driverinformation {
    id: number;
    code: string;
    name: string;
    dateOfBirth: Date;
    dateOfJoining: Date;
    licenseNo: string;
    issueDate: Date;
    validUpto: Date;
    citizenshipNo: string;
    csnIssueDate: Date;
    unitNo: string;
    remarks: string;
    status: string;
    orgId: number;
    orgname: string;
    suborgname: string;
    subOrgId: number;
    currentUserId: number;
    state: boolean;
    lang: string;
}